Replace old MHA by new
